refactor: converted all GraphQL filenames to use Pascal case
