+++ src/apollo/groups/fragments/GroupAvatar.fragment.ts
