+++ src/apollo/groups/fragments/GroupCard.fragment.ts
@@ -1,5 +1,5 @@
 import { gql } from "@apollo/client";
-import GROUP_AVATAR_FRAGMENT from "./group-avatar.fragment";
+import GROUP_AVATAR_FRAGMENT from "./GroupAvatar.fragment";
 
 const GROUP_CARD_FRAGMENT = gql`
   fragment GroupCard on Group {
