+++ src/apollo/groups/fragments/GroupProfileCard.fragment.ts
