+++ src/apollo/groups/mutations/LeaveGroup.mutation.ts
