+++ src/apollo/groups/queries/MemberRequests.query.ts
@@ -1,5 +1,5 @@
 import { gql } from "@apollo/client";
-import USER_AVATAR_FRAGMENT from "../../users/fragments/user-avatar.fragment";
+import USER_AVATAR_FRAGMENT from "../../users/fragments/UserAvatar.fragment";
 
 const MEMBER_REQUESTS_QUERY = gql`
   query MemberRequests($groupId: Int!) {
