+++ src/apollo/groups/queries/groups.query.ts
@@ -1,5 +1,5 @@
 import { gql } from "@apollo/client";
-import GROUP_CARD_FRAGMENT from "../fragments/group-card.fragment";
+import GROUP_CARD_FRAGMENT from "../fragments/GroupCard.fragment";
 
 const GROUPS_QUERY = gql`
   query Groups {
