+++ src/apollo/images/fragments/AttachedImage.fragment.ts
