+++ src/apollo/images/mutations/DeleteImage.mutation.ts
