+++ src/apollo/posts/mutations/CreatePost.mutation.ts
@@ -2,8 +2,8 @@ import { gql } from "@apollo/client";
 import { ApiRoutes, HttpMethod } from "../../../constants/common.constants";
 import { Image } from "../../../types/generated.types";
 import { multiPartRequest } from "../../../utils/common.utils";
-import GROUP_AVATAR_FRAGMENT from "../../groups/fragments/group-avatar.fragment";
-import USER_AVATAR_FRAGMENT from "../../users/fragments/user-avatar.fragment";
+import GROUP_AVATAR_FRAGMENT from "../../groups/fragments/GroupAvatar.fragment";
+import USER_AVATAR_FRAGMENT from "../../users/fragments/UserAvatar.fragment";
 
 const CREATE_POST_MUTATION = gql`
   mutation CreatePost($postData: PostInput!) {
