+++ src/apollo/users/fragments/UserAvatar.fragment.ts
