+++ src/apollo/users/fragments/UserProfileCard.fragment.ts
@@ -1,5 +1,5 @@
 import { gql } from "@apollo/client";
-import USER_AVATAR_FRAGMENT from "./user-avatar.fragment";
+import USER_AVATAR_FRAGMENT from "./UserAvatar.fragment";
 
 const USER_PROFILE_CARD_FRAGMENT = gql`
   fragment UserProfileCard on User {
