+++ src/apollo/users/queries/user.query.ts
@@ -1,6 +1,6 @@
 import { gql } from "@apollo/client";
-import POST_CARD_FRAGMENT from "../../posts/fragments/post-card.fragment";
-import USER_PROFILE_CARD_FRAGMENT from "../fragments/user-profile-card.fragment";
+import POST_CARD_FRAGMENT from "../../posts/fragments/PostCard.fragment";
+import USER_PROFILE_CARD_FRAGMENT from "../fragments/UserProfileCard.fragment";
 
 const USER_QUERY = gql`
   query User($name: String) {
