+++ src/components/Groups/GroupForm.tsx
@@ -11,8 +11,8 @@ import { Form, Formik, FormikHelpers } from "formik";
 import produce from "immer";
 import { useState } from "react";
 import { toastVar } from "../../apollo/cache";
-import { uploadGroupCoverPhoto } from "../../apollo/groups/mutations/create-group.mutation";
-import GROUPS_QUERY from "../../apollo/groups/queries/groups.query";
+import { uploadGroupCoverPhoto } from "../../apollo/groups/mutations/CreateGroup.mutation";
+import GROUPS_QUERY from "../../apollo/groups/queries/Groups.query";
 import Flex from "../../components/Shared/Flex";
 import { TextField } from "../../components/Shared/TextField";
 import { FieldNames } from "../../constants/common.constants";
