+++ src/components/Navigation/NavDrawer.tsx
@@ -18,7 +18,7 @@ import { styled, SxProps } from "@mui/material/styles";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 import { isLoggedInVar, isNavDrawerOpenVar } from "../../apollo/cache";
-import ME_QUERY from "../../apollo/users/queries/me.query";
+import ME_QUERY from "../../apollo/users/queries/Me.query";
 import { NavigationPaths } from "../../constants/common.constants";
 import { useTranslate } from "../../hooks/common.hooks";
 import { useLogOutMutation, useMeQuery } from "../../types/generated.types";
