+++ src/components/Navigation/TopNavDropdown.tsx
@@ -5,7 +5,7 @@ import {
   isLoggedInVar,
   isRefreshingTokenVar,
 } from "../../apollo/cache";
-import ME_QUERY from "../../apollo/users/queries/me.query";
+import ME_QUERY from "../../apollo/users/queries/Me.query";
 import {
   NavigationPaths,
   ResourceNames,
